requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==2.0.6
dload==0.6
idna==2.10
numpy==1.21.2
oauthlib==3.1.0
PyMySQL==1.0.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.25.1
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.15.0
soupsieve==2.2.1
tweepy==3.10.0
urllib3==1.26.4


Pillow~=8.4.0